+++ pollux/lib/core/src/main/scala/io/iohk/atala/pollux/core/service/CredentialServiceImpl.scala
@@ -34,6 +34,9 @@ import java.security.SecureRandom
 import java.security.spec.ECGenParameterSpec
 import java.time.Instant
 import java.util.UUID
+import io.iohk.atala.castor.core.model.did.CanonicalPrismDID
+import io.iohk.atala.mercury.model.AttachmentDescriptor
+import io.iohk.atala.pollux.core.model.CredentialOfferAttachment
 
 object CredentialServiceImpl {
   val layer: URLayer[IrisServiceStub & CredentialRepository[Task], CredentialService] =
@@ -67,18 +70,26 @@ private class CredentialServiceImpl(
   }
 
   override def createIssueCredentialRecord(
-      pairwiseDID: DidId,
+      pairwiseIssuerDID: DidId,
+      pairwiseHolderDID: DidId,
       thid: UUID,
       subjectId: String,
       schemaId: Option[String],
       claims: Map[String, String],
       validityPeriod: Option[Double],
       automaticIssuance: Option[Boolean],
-      awaitConfirmation: Option[Boolean]
+      awaitConfirmation: Option[Boolean],
+      issuingDID: Option[CanonicalPrismDID]
   ): IO[CredentialServiceError, IssueCredentialRecord] = {
     for {
       _ <- if (DidValidator.supportedDid(subjectId)) ZIO.unit else ZIO.fail(UnsupportedDidFormat(subjectId))
-      offer <- ZIO.succeed(createDidCommOfferCredential(pairwiseDID, claims, thid, subjectId))
+      offer = createDidCommOfferCredential(
+        pairwiseIssuerDID = pairwiseIssuerDID,
+        pairwiseHolderDID = pairwiseHolderDID,
+        claims = claims,
+        thid = thid,
+        subjectId = subjectId
+      )
       record <- ZIO.succeed(
         IssueCredentialRecord(
           id = UUID.randomUUID(),
@@ -96,7 +107,8 @@ private class CredentialServiceImpl(
           offerCredentialData = Some(offer),
           requestCredentialData = None,
           issueCredentialData = None,
-          issuedCredentialRaw = None
+          issuedCredentialRaw = None,
+          issuingDID = issuingDID
         )
       )
       count <- credentialRepository
@@ -123,6 +135,17 @@ private class CredentialServiceImpl(
       offer: OfferCredential
   ): IO[CredentialServiceError, IssueCredentialRecord] = {
     for {
+      // TODO: align with the standard (ATL-3507)
+      offerAttachment <- offer.attachments.headOption
+        .map(_.data.asJson)
+        .fold(ZIO.fail(CredentialServiceError.UnexpectedError("An attachment is expected in CredentialOffer"))) {
+          json =>
+            ZIO
+              .fromTry(json.hcursor.downField("json").as[CredentialOfferAttachment].toTry)
+              .mapError(e =>
+                CredentialServiceError.UnexpectedError(s"Unexpected CredentialOffer attachment format: ${e.toString()}")
+              )
+        }
       record <- ZIO.succeed(
         IssueCredentialRecord(
           id = UUID.randomUUID(),
@@ -131,7 +154,7 @@ private class CredentialServiceImpl(
           thid = UUID.fromString(offer.thid.getOrElse(offer.id)),
           schemaId = None,
           role = Role.Holder,
-          subjectId = offer.to.value,
+          subjectId = offerAttachment.subjectId,
           validityPeriod = None,
           automaticIssuance = None,
           awaitConfirmation = None,
@@ -140,7 +163,8 @@ private class CredentialServiceImpl(
           offerCredentialData = Some(offer),
           requestCredentialData = None,
           issueCredentialData = None,
-          issuedCredentialRaw = None
+          issuedCredentialRaw = None,
+          issuingDID = None
         )
       )
       count <- credentialRepository
@@ -153,7 +177,7 @@ private class CredentialServiceImpl(
     } yield record
   }
 
-  override def acceptCredentialOffer(recordId: UUID): IO[CredentialServiceError, Option[IssueCredentialRecord]] = {
+  override def acceptCredentialOffer(recordId: UUID): IO[CredentialServiceError, IssueCredentialRecord] = {
     for {
       record <- getRecordWithState(recordId, ProtocolState.OfferReceived)
       offer <- ZIO
@@ -169,12 +193,16 @@ private class CredentialServiceImpl(
       record <- credentialRepository
         .getIssueCredentialRecord(record.id)
         .mapError(RepositoryError.apply)
+        .flatMap {
+          case None        => ZIO.fail(RecordIdNotFound(recordId))
+          case Some(value) => ZIO.succeed(value)
+        }
     } yield record
   }
 
   override def receiveCredentialRequest(
       request: RequestCredential
-  ): IO[CredentialServiceError, Option[IssueCredentialRecord]] = {
+  ): IO[CredentialServiceError, IssueCredentialRecord] = {
     for {
       record <- getRecordFromThreadIdWithState(request.thid, ProtocolState.OfferPending, ProtocolState.OfferSent)
       _ <- credentialRepository
@@ -187,10 +215,11 @@ private class CredentialServiceImpl(
       record <- credentialRepository
         .getIssueCredentialRecord(record.id)
         .mapError(RepositoryError.apply)
+        .someOrFail(RecordIdNotFound(record.id))
     } yield record
   }
 
-  override def acceptCredentialRequest(recordId: UUID): IO[CredentialServiceError, Option[IssueCredentialRecord]] = {
+  override def acceptCredentialRequest(recordId: UUID): IO[CredentialServiceError, IssueCredentialRecord] = {
     for {
       record <- getRecordWithState(recordId, ProtocolState.RequestReceived)
       request <- ZIO
@@ -206,12 +235,13 @@ private class CredentialServiceImpl(
       record <- credentialRepository
         .getIssueCredentialRecord(record.id)
         .mapError(RepositoryError.apply)
+        .someOrFail(RecordIdNotFound(record.id))
     } yield record
   }
 
   override def receiveCredentialIssue(
       issue: IssueCredential
-  ): IO[CredentialServiceError, Option[IssueCredentialRecord]] = {
+  ): IO[CredentialServiceError, IssueCredentialRecord] = {
     val rawIssuedCredential = issue.attachments.map(_.data.asJson.noSpaces).headOption.getOrElse("???") // TODO
     for {
       record <- getRecordFromThreadIdWithState(issue.thid, ProtocolState.RequestPending, ProtocolState.RequestSent)
@@ -225,17 +255,18 @@ private class CredentialServiceImpl(
       record <- credentialRepository
         .getIssueCredentialRecord(record.id)
         .mapError(RepositoryError.apply)
+        .someOrFail(RecordIdNotFound(record.id))
     } yield record
   }
 
-  override def markOfferSent(recordId: UUID): IO[CredentialServiceError, Option[IssueCredentialRecord]] =
+  override def markOfferSent(recordId: UUID): IO[CredentialServiceError, IssueCredentialRecord] =
     updateCredentialRecordProtocolState(
       recordId,
       IssueCredentialRecord.ProtocolState.OfferPending,
       IssueCredentialRecord.ProtocolState.OfferSent
     )
 
-  override def markRequestSent(recordId: UUID): IO[CredentialServiceError, Option[IssueCredentialRecord]] =
+  override def markRequestSent(recordId: UUID): IO[CredentialServiceError, IssueCredentialRecord] =
     updateCredentialRecordProtocolState(
       recordId,
       IssueCredentialRecord.ProtocolState.RequestPending,
@@ -245,7 +276,7 @@ private class CredentialServiceImpl(
   override def markCredentialGenerated(
       recordId: UUID,
       issueCredential: IssueCredential
-  ): IO[CredentialServiceError, Option[IssueCredentialRecord]] = {
+  ): IO[CredentialServiceError, IssueCredentialRecord] = {
     for {
       record <- getRecordWithState(recordId, ProtocolState.CredentialPending)
       count <- credentialRepository
@@ -261,10 +292,15 @@ private class CredentialServiceImpl(
       record <- credentialRepository
         .getIssueCredentialRecord(recordId)
         .mapError(RepositoryError.apply)
+        .flatMap {
+          case None        => ZIO.fail(RecordIdNotFound(recordId))
+          case Some(value) => ZIO.succeed(value)
+        }
+
     } yield record
   }
 
-  override def markCredentialSent(recordId: UUID): IO[CredentialServiceError, Option[IssueCredentialRecord]] =
+  override def markCredentialSent(recordId: UUID): IO[CredentialServiceError, IssueCredentialRecord] =
     updateCredentialRecordProtocolState(
       recordId,
       IssueCredentialRecord.ProtocolState.CredentialGenerated,
@@ -273,7 +309,7 @@ private class CredentialServiceImpl(
 
   override def markCredentialPublicationPending(
       recordId: UUID
-  ): IO[CredentialServiceError, Option[IssueCredentialRecord]] =
+  ): IO[CredentialServiceError, IssueCredentialRecord] =
     updateCredentialRecordPublicationState(
       recordId,
       None,
@@ -282,14 +318,14 @@ private class CredentialServiceImpl(
 
   override def markCredentialPublicationQueued(
       recordId: UUID
-  ): IO[CredentialServiceError, Option[IssueCredentialRecord]] =
+  ): IO[CredentialServiceError, IssueCredentialRecord] =
     updateCredentialRecordPublicationState(
       recordId,
       Some(IssueCredentialRecord.PublicationState.PublicationPending),
       Some(IssueCredentialRecord.PublicationState.PublicationQueued)
     )
 
-  override def markCredentialPublished(recordId: UUID): IO[CredentialServiceError, Option[IssueCredentialRecord]] =
+  override def markCredentialPublished(recordId: UUID): IO[CredentialServiceError, IssueCredentialRecord] =
     updateCredentialRecordPublicationState(
       recordId,
       Some(IssueCredentialRecord.PublicationState.PublicationQueued),
@@ -337,7 +373,8 @@ private class CredentialServiceImpl(
   }
 
   private[this] def createDidCommOfferCredential(
-      pairwiseDID: DidId,
+      pairwiseIssuerDID: DidId,
+      pairwiseHolderDID: DidId,
       claims: Map[String, String],
       thid: UUID,
       subjectId: String
@@ -348,9 +385,14 @@ private class CredentialServiceImpl(
 
     OfferCredential(
       body = body,
-      attachments = Seq(),
-      from = pairwiseDID,
-      to = DidId(subjectId),
+      // TODO: align with the standard (ATL-3507)
+      attachments = Seq(
+        AttachmentDescriptor.buildJsonAttachment(
+          payload = CredentialOfferAttachment(subjectId = subjectId)
+        )
+      ),
+      from = pairwiseIssuerDID,
+      to = pairwiseHolderDID,
       thid = Some(thid.toString())
     )
   }
@@ -385,41 +427,110 @@ private class CredentialServiceImpl(
     )
   }
 
+  /** this is an auxiliary function.
+    *
+    * @note
+    *   Between updating and getting the CredentialRecord back the CredentialRecord can be updated by other operations
+    *   in the middle.
+    *
+    * TODO: this should be improved to behave exactly like atomic operation.
+    */
   private[this] def updateCredentialRecordProtocolState(
       id: UUID,
       from: IssueCredentialRecord.ProtocolState,
       to: IssueCredentialRecord.ProtocolState
-  ): IO[CredentialServiceError, Option[IssueCredentialRecord]] = {
+  ): IO[CredentialServiceError, IssueCredentialRecord] = {
     for {
-      outcome <- credentialRepository
-        .updateCredentialRecordProtocolState(id, from, to)
-        .flatMap {
-          case 1 => ZIO.succeed(())
-          case n => ZIO.fail(UnexpectedException(s"Invalid row count result: $n"))
-        }
-        .mapError(RepositoryError.apply)
       record <- credentialRepository
-        .getIssueCredentialRecord(id)
+        .updateCredentialRecordProtocolState(id, from, to)
         .mapError(RepositoryError.apply)
+        .flatMap {
+          case 0 =>
+            credentialRepository
+              .getIssueCredentialRecord(id)
+              .mapError(RepositoryError.apply)
+              .flatMap {
+                case None => ZIO.fail(RecordIdNotFound(id))
+                case Some(record) if record.protocolState == to => // Not update by is alredy on the desirable state
+                  ZIO.succeed(record)
+                case Some(record) =>
+                  ZIO.fail(
+                    OperationNotExecuted(
+                      id,
+                      s"CredentialRecord was not updated because have the ProtocolState ${record.protocolState}"
+                    )
+                  )
+              }
+          case 1 =>
+            credentialRepository
+              .getIssueCredentialRecord(id)
+              .mapError(RepositoryError.apply)
+              .flatMap {
+                case None => ZIO.fail(RecordIdNotFound(id))
+                case Some(record) =>
+                  ZIO
+                    .logError(
+                      s"The CredentialRecord ($id) is expected to be on the ProtocolState '$to' after updating it"
+                    )
+                    .when(record.protocolState != to)
+                    .as(record)
+              }
+          case n => ZIO.fail(UnexpectedError(s"Invalid row count result: $n"))
+        }
     } yield record
   }
 
+  /** this is an auxiliary function.
+    *
+    * @note
+    *   Between updating and getting the CredentialRecord back the CredentialRecord can be updated by other operations
+    *   in the middle.
+    *
+    * TODO: this should be improved to behave exactly like atomic operation.
+    */
   private[this] def updateCredentialRecordPublicationState(
       id: UUID,
       from: Option[IssueCredentialRecord.PublicationState],
       to: Option[IssueCredentialRecord.PublicationState]
-  ): IO[CredentialServiceError, Option[IssueCredentialRecord]] = {
+  ): IO[CredentialServiceError, IssueCredentialRecord] = {
     for {
-      outcome <- credentialRepository
-        .updateCredentialRecordPublicationState(id, from, to)
-        .flatMap {
-          case 1 => ZIO.succeed(())
-          case n => ZIO.fail(UnexpectedException(s"Invalid row count result: $n"))
-        }
-        .mapError(RepositoryError.apply)
       record <- credentialRepository
-        .getIssueCredentialRecord(id)
+        .updateCredentialRecordPublicationState(id, from, to)
         .mapError(RepositoryError.apply)
+        .flatMap {
+          case 0 =>
+            credentialRepository
+              .getIssueCredentialRecord(id)
+              .mapError(RepositoryError.apply)
+              .flatMap {
+                case None => ZIO.fail(RecordIdNotFound(id))
+                case Some(record) if record.publicationState == to => // Not update by is alredy on the desirable state
+                  ZIO.succeed(record)
+                case Some(record) =>
+                  ZIO.fail(
+                    OperationNotExecuted(
+                      id,
+                      s"CredentialRecord was not updated because have the PublicationState ${record.publicationState}"
+                    )
+                  )
+              }
+          case 1 =>
+            credentialRepository
+              .getIssueCredentialRecord(id)
+              .mapError(RepositoryError.apply)
+              .flatMap {
+                case None => ZIO.fail(RecordIdNotFound(id))
+                case Some(record) =>
+                  {
+                    if (record.publicationState == to) (ZIO.unit)
+                    else
+                      ZIO.logError(
+                        s"The CredentialRecord ($id) is expected to be on the PublicationState '$to' after updating it"
+                      ) // The expectation is for the record to still be on the state we (just) updated to
+                  } *> ZIO.succeed(record)
+              }
+          case n => ZIO.fail(UnexpectedError(s"Invalid row count result: $n"))
+        }
     } yield record
   }