+++ prism-agent/service/wallet-api/src/main/scala/io/iohk/atala/agent/walletapi/model/error/GetKeyError.scala
@@ -1,0 +1,8 @@
+package io.iohk.atala.agent.walletapi.model.error
+
+sealed trait GetKeyError
+
+object GetKeyError {
+  final case class WalletStorageError(cause: Throwable) extends GetKeyError
+  final case class KeyConstructionError(cause: Throwable) extends GetKeyError
+}